+++ app/Notifications/Component/ComponentStatusChangedNotification.php
@@ -95,6 +95,8 @@ class ComponentStatusChangedNotification extends Notification
                 'unsubscribeUrl'         => cachet_route('subscribe.unsubscribe', $notifiable->verify_code),
                 'manageSubscriptionText' => trans('cachet.subscriber.manage_subscription'),
                 'manageSubscriptionUrl'  => $manageUrl,
+                'statuspageText'         => trans('cachet.subscriber.statuspage'),
+                'statuspageUrl'          => cachet_route('status-page'), 
             ]);
     }