+++ app/Notifications/IncidentUpdate/IncidentUpdatedNotification.php
@@ -91,6 +91,8 @@ class IncidentUpdatedNotification extends Notification
                 'unsubscribeUrl'         => cachet_route('subscribe.unsubscribe', $notifiable->verify_code),
                 'manageSubscriptionText' => trans('cachet.subscriber.manage_subscription'),
                 'manageSubscriptionUrl'  => $manageUrl,
+                'statuspageText'         => trans('cachet.subscriber.statuspage'),
+                'statuspageUrl'          => cachet_route('status-page'), 
             ]);
     }