+++ app/Notifications/Schedule/NewScheduleNotification.php
@@ -85,6 +85,8 @@ class NewScheduleNotification extends Notification implements ShouldQueue
                 'unsubscribeUrl'         => cachet_route('subscribe.unsubscribe', $notifiable->verify_code),
                 'manageSubscriptionText' => trans('cachet.subscriber.manage_subscription'),
                 'manageSubscriptionUrl'  => $manageUrl,
+                'statuspageText'         => trans('cachet.subscriber.statuspage'),
+                'statuspageUrl'          => cachet_route('status-page'), 
             ]);
     }