+++ resources/lang/fr/cachet.php
@@ -93,10 +93,11 @@ return [
             ],
             'component' => [
                 'subject'       => 'Mise à jour du statut d\'un composant',
-                'text'          => 'Le statut du composant :component_name a été mis à jour. Il est maintenant :component_human_status.\nMerci, :app_name',
-                'html'          => '<p>Le statut du composant :component_name a été mis à jour. Il est maintenant :component_human_status.</p><p>Merci, :app_name</p>',
+                'text'          => 'Le statut du composant :component_name a été mis à jour. Il est maintenant en l\'état: \'<b>:component_human_status</b>\'.\nMerci, :app_name',
+                'html'          => '<p>Le statut du composant :component_name a été mis à jour. Il est maintenant en l\'état: \'<b>:component_human_status</b>\'.</p><p>Merci, :app_name</p>',
                 'tooltip-title' => 'S\'abonner aux notifications pour :component_name.',
             ],
+            'statuspage'         => 'Page de statut',            ],
         ],
     ],